More general arrow_box css
The css for the arrow box had an uneven padding declaration, now it
doesn't. The padding's purpose was to center the textbox it contains,
while ignoring the textbox's preceding prompt character. Since the arrow
box will soon have to accept other contents, the hack was removed and
was replaced by a better hack.